+++ src/WEDCore/WED_Buffer.h
@@ -76,7 +76,7 @@ public:
 
 			void	ResetRead(void);				// Resets reading to the beginning.
 			void *	AllocContiguous(int len);		// Allocates a chunk of memory that is not split across buffers.
-			void	GetWritePos(long& a, long& b);	// Get and set arbitrary positions as two ints.
+			void	GetWritePos(intptr_t& a, intptr_t& b);	// Get and set arbitrary positions as two ints.
 			void	SetReadPos(uintptr_t a, int b);
 
 private:
